refatoraçao de testes unitarios, implementa testes de integracao, buscas e filtros avancados e outros ajustes diversos

+++ src/main/java/com/example/federacao_develop/service/PartidaService.java
@@ -1,18 +1,17 @@
 package com.example.federacao_develop.service;
 
 import com.example.federacao_develop.dto.PartidaDTO;
+import com.example.federacao_develop.exception.BusinessException;
+import com.example.federacao_develop.exception.MensagemExceptionEnum;
+import com.example.federacao_develop.exception.NotFoundException;
 import com.example.federacao_develop.mapper.PartidaMapper;
-import com.example.federacao_develop.model.Clube;
-import com.example.federacao_develop.model.Estadio;
 import com.example.federacao_develop.model.Partida;
 import com.example.federacao_develop.repository.ClubeRepository;
 import com.example.federacao_develop.repository.EstadioRepository;
 import com.example.federacao_develop.repository.PartidaRepository;
 import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 
 @Service
 public class PartidaService {
@@ -30,15 +29,14 @@ public class PartidaService {
         this.partidaMapper = partidaMapper;
     }
 
-    public List<PartidaDTO> findAll() {
-        return partidaRepository.findAll().stream()
-                .map(partidaMapper::toDTO)
-                .collect(Collectors.toList());
+    public Page<PartidaDTO> findPartidasFilter(Integer clubeId, Integer estadioId, Pageable pageable) {
+        return partidaRepository.findPartidasFilter(clubeId, estadioId, pageable)
+                .map(partidaMapper::toDTO);
     }
 
     public PartidaDTO findById(Integer id) {
         Partida partida = partidaRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Partida não encontrada"));
+                .orElseThrow(() -> new NotFoundException(MensagemExceptionEnum.PARTIDA_NAO_ENCONTRADA));
         return partidaMapper.toDTO(partida);
     }
 
@@ -47,15 +45,21 @@ public class PartidaService {
 
         if (dto.getClubeMandanteId() != null) {
             entity.setClubeMandante(clubeRepository.findById(dto.getClubeMandanteId())
-                    .orElseThrow(() -> new RuntimeException("Clube mandante não encontrado")));
+                    .orElseThrow(() -> new NotFoundException(MensagemExceptionEnum.CLUBE_MANDANTE_NAO_ENCONTRADO)));
+        } else {
+            throw new BusinessException(MensagemExceptionEnum.CLUBE_MANDANTE_NAO_ENCONTRADO);
         }
         if (dto.getClubeVisitanteId() != null) {
             entity.setClubeVisitante(clubeRepository.findById(dto.getClubeVisitanteId())
-                    .orElseThrow(() -> new RuntimeException("Clube visitante não encontrado")));
+                    .orElseThrow(() -> new NotFoundException(MensagemExceptionEnum.CLUBE_VISITANTE_NAO_ENCONTRADO)));
+        } else {
+            throw new BusinessException(MensagemExceptionEnum.CLUBE_VISITANTE_NAO_ENCONTRADO);
         }
         if (dto.getEstadio() != null && dto.getEstadio().getEstadioId() != null) {
             entity.setEstadio(estadioRepository.findById(dto.getEstadio().getEstadioId())
-                    .orElseThrow(() -> new RuntimeException("Estádio não encontrado")));
+                    .orElseThrow(() -> new NotFoundException(MensagemExceptionEnum.ESTADIO_NAO_ENCONTRADO)));
+        } else {
+            throw new BusinessException(MensagemExceptionEnum.ESTADIO_NAO_ENCONTRADO);
         }
 
         return partidaMapper.toDTO(partidaRepository.save(entity));
@@ -63,29 +67,33 @@ public class PartidaService {
 
     public PartidaDTO update(Integer id, PartidaDTO dto) {
         Partida entity = partidaRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Partida não encontrada"));
+                .orElseThrow(() -> new NotFoundException(MensagemExceptionEnum.PARTIDA_NAO_ENCONTRADA));
 
         if (dto.getClubeMandanteId() != null) {
             entity.setClubeMandante(clubeRepository.findById(dto.getClubeMandanteId())
-                    .orElseThrow(() -> new RuntimeException("Clube mandante não encontrado")));
+                    .orElseThrow(() -> new NotFoundException(MensagemExceptionEnum.CLUBE_MANDANTE_NAO_ENCONTRADO)));
         }
         if (dto.getClubeVisitanteId() != null) {
             entity.setClubeVisitante(clubeRepository.findById(dto.getClubeVisitanteId())
-                    .orElseThrow(() -> new RuntimeException("Clube visitante não encontrado")));
+                    .orElseThrow(() -> new NotFoundException(MensagemExceptionEnum.CLUBE_VISITANTE_NAO_ENCONTRADO)));
         }
+        if (dto.getEstadio() != null && dto.getEstadio().getEstadioId() != null) {
+            entity.setEstadio(estadioRepository.findById(dto.getEstadio().getEstadioId())
+                    .orElseThrow(() -> new NotFoundException(MensagemExceptionEnum.ESTADIO_NAO_ENCONTRADO)));
+        }
+
         entity.setGolsMandante(dto.getGolsMandante());
         entity.setGolsVisitante(dto.getGolsVisitante());
         entity.setDataDaPartida(dto.getDataDaPartida());
-
-        if (dto.getEstadio() != null && dto.getEstadio().getEstadioId() != null) {
-            entity.setEstadio(estadioRepository.findById(dto.getEstadio().getEstadioId())
-                    .orElseThrow(() -> new RuntimeException("Estádio não encontrado")));
-        }
 
         return partidaMapper.toDTO(partidaRepository.save(entity));
     }
 
     public void delete(Integer id) {
+        boolean exists = partidaRepository.existsById(id);
+        if (!exists) {
+            throw new NotFoundException(MensagemExceptionEnum.PARTIDA_NAO_ENCONTRADA);
+        }
         partidaRepository.deleteById(id);
     }
 }
